+++ packages/networks/src/networks/xdai.ts
@@ -9,6 +9,7 @@ export const xdai: NetworkConfig = {
   id: 100,
   name: 'Gnosis Chain',
   blockTime: 5000,
+  chain: 'xdai',
   requiredConfirmations: 12,
   subgraph: {
     endpoint: 'https://api.thegraph.com/subgraphs/name/unlock-protocol/xdai',
@@ -35,6 +36,7 @@ export const xdai: NetworkConfig = {
     name: 'DAI',
     symbol: 'DAI',
     decimals: 18,
+    coingecko: 'xdai',
   },
   startBlock: 19338700,
   previousDeploys: [
@@ -47,13 +49,18 @@ export const xdai: NetworkConfig = {
     'EVM compatible network whose base currency is a stable coin. Cheaper transaction cost.',
   isTestNetwork: false,
   teamMultisig: '0xfAC611a5b5a578628C28F77cEBDDB8C6159Ae79D',
-  // 0xddafbb505ad214d7b80b1f830fccc89b60fb7a83
   tokens: [
     {
       name: 'USD Coin',
       symbol: 'USDC',
       decimals: 6,
       address: '0xddafbb505ad214d7b80b1f830fccc89b60fb7a83',
+    },
+    {
+      name: 'Tether USD',
+      decimals: 6,
+      address: '0x4ecaba5870353805a9f068101a40e0f32ed605c6',
+      symbol: 'USDT',
     },
   ],
 }