+++ packages/unlock-js/src/index.ts
@@ -8,3 +8,4 @@ export {
   LocksmithServiceConfiguration,
 } from './locksmithService'
 export { SubgraphService } from './subgraph'
+export { UniswapService } from './uniswapService'