+++ packages/unlock-js/CHANGELOG.md
@@ -1,5 +1,45 @@
 # Changes
 
+# 0.32.5
+
+- add `updateRefundPenalty` in walletService
+- add `freeTrialLength` in web3Service
+- add `refundPenaltyBasisPoints` in web3Service
+
+# 0.32.4
+
+- Using `transactionOptions` for `createLock`
+
+# 0.32.3
+
+- Pass network property on locks and keys returned from subgraph.
+
+# 0.32.1
+
+- Change parameter name for `updateLockSymbol`
+
+# 0.32.0
+
+Allow passing creator in the createLock method and improve typing.
+
+# 0.31.6
+
+- add `addLockManager` in walletService
+- add `renounceLockManager` in walletService
+
+# 0.31.5
+
+- Added suppport for `runEstimate`
+
+# 0.31.4
+
+- Add check to ensure lock is deployed from unlock factory contract in getLock method
+
+# 0.31.2
+
+- Add uniswap service
+- Add consultUniswap method to web3 service for getting price of a token using uniswap v3 pools
+
 # 0.31.1
 
 - add `totalKeys` to subgraph